refactor(frontend): split SettingsModal inline panels into settings/*
Phase 5.1 — no-behavior/no-visual split. SettingsModal was already ~80% a router; this
extracts the 3 remaining large inline sections, leaving the modal shell + sidebar + the
{#if activeSection} routing untouched (2932 -> 1333):
- settings/UserProfileSettings.svelte (482) — profile/password form; owns its form state,
  updateProfile/updatePassword, the 'profile' dirty key, inits from authStore; renders
  LanguageSettings + SecuritySettings.
- settings/SystemStatisticsPanel.svelte (967, mostly scoped CSS) — stats grid + recent-tasks
  table; props (stats/loading/search status), events refresh/openDetails; owns GPU-stepper
  state. Modal keeps loadStats + ProcessingDetailsModal.
- settings/AdminTaskHealthPanel.svelte (614) — recovery actions + stuck-tasks/inconsistent-
  files tables; self-loads health; dispatches requestConfirm to the modal's ConfirmationModal.

Verified: svelte-check 0/0 (852 files), build OK, 52 unit tests, settings E2E 3/3 green
(modal opens, all sections incl. the moved Profile render, closes). Guardrail selectors
(.settings-modal[role=dialog], .settings-sidebar, .nav-item, .section-title,
.modal-close-button) intact.
